+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.915](https://github.com/parca-dev/parca/compare/@parca/web@0.16.914...@parca/web@0.16.915) (2025-08-06)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.914](https://github.com/parca-dev/parca/compare/@parca/web@0.16.913...@parca/web@0.16.914) (2025-08-06)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.913](https://github.com/parca-dev/parca/compare/@parca/web@0.16.912...@parca/web@0.16.913) (2025-08-06)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.912](https://github.com/parca-dev/parca/compare/@parca/web@0.16.911...@parca/web@0.16.912) (2025-08-04)
 
 **Note:** Version bump only for package @parca/web